+++ handlers/transaction_handlers.js
@@ -1,7 +1,6 @@
 const {fetchMetadataAtHeight, injectMetadata} = require('../utils/setup');
-const { CalcFee } = require('@substrate/calc');
+const {createCalcFee} = require("../utils/calc");
 const transactionMappers = require('../mappers/transaction/transaction_mappers');
-const eventMappers = require('../mappers/event/event_mappers');
 
 /**
  * Get signed transactions by height
@@ -17,9 +16,11 @@ const getByHeight = async (api, call, context = {}) => {
   const resp = await api.rpc.chain.getBlock(blockHash);
   const rawBlockAt = resp.block;
 
-  const rawTimestampAt = await api.query.timestamp.now.at(blockHash);
-  const rawEventsAt = await api.query.system.events.at(blockHash);
-  const calcFee = await createCalcFee(api, rawBlockAt);
+  const [rawTimestampAt, rawEventsAt, calcFee] = await Promise.all([
+    api.query.timestamp.now.at(blockHash),
+    api.query.system.events.at(blockHash),
+    createCalcFee(api, rawBlockAt),
+  ]);
 
   const transactions = [];
 
@@ -28,61 +29,11 @@ const getByHeight = async (api, call, context = {}) => {
       return
     }
     const rawEventsForExtrinsic = rawEventsAt.filter((ev) => ev.phase.isApplyExtrinsic && ev.phase.asApplyExtrinsic.toNumber() === index);
-    const completedRawEvent =  rawEventsForExtrinsic.find(({ event }) =>
-        event.section == 'system' && (event.method === 'ExtrinsicSuccess' || event.method === 'ExtrinsicFailure')
-    );
-
-    var fee;
-    if (calcFee && completedRawEvent) {
-      const completedEvent =  eventMappers.toPb(completedRawEvent);
-      const weightInfo = completedEvent.data.find(({name}) => name == "DispatchInfo")
-      const weight = JSON.parse(weightInfo.value).weight;
-      const partialFee = calcFee.calc_fee(
-        BigInt(weight.toString()),
-        rawExtrinsic.encodedLength
-      );
-
-      fee = partialFee // todo add tip for full fee
-    }
-
-      transactions.push(transactionMappers.toPb(index, rawExtrinsic, rawTimestampAt, rawEventsForExtrinsic));
+    transactions.push(transactionMappers.toPb(index, rawExtrinsic, rawTimestampAt, rawEventsForExtrinsic, calcFee));
   });
 
   return {transactions};
 };
-
-
-const createCalcFee = async(api, block) => {
-  const { parentHash } = block.header;
-  const perByte = api.consts.transactionPayment.transactionByteFee;
-  const extrinsicBaseWeight = api.consts.system.extrinsicBaseWeight;
-  const multiplier = await api.query.transactionPayment.nextFeeMultiplier.at(
-    parentHash
-  );
-  const version = await api.rpc.state.getRuntimeVersion(parentHash); // TODO pass this?
-  const specName = version.specName.toString();
-  const specVersion = version.specVersion.toNumber();
-  const coefficients = api.consts.transactionPayment.weightToFee.map(
-    (c) => {
-      return {
-        coeffInteger: c.coeffInteger.toString(),
-        coeffFrac: c.coeffFrac,
-        degree: c.degree,
-        negative: c.negative,
-      };
-    }
-  );
-
-  return CalcFee.from_params(
-      coefficients,
-      BigInt(extrinsicBaseWeight.toString()),
-      multiplier.toString(),
-      perByte.toString(),
-      specName,
-      specVersion
-    );
-};
-
 
 module.exports = {
   getByHeight,
