+++ mappers/block/block_mappers.js
@@ -1,6 +1,6 @@
 const transactionMappers = require('../transaction/transaction_mappers');
 
-const toPb = (blockHash, rawBlock, rawTimestamp, rawEvents) => {
+const toPb = (blockHash, rawBlock, rawTimestamp, rawEvents, calcFee) => {
   return {
     block: {
       blockHash: blockHash.toString(),
@@ -13,7 +13,7 @@ const toPb = (blockHash, rawBlock, rawTimestamp, rawEvents) => {
       },
       extrinsics: rawBlock.extrinsics.map((rawExtrinsic, index) => {
         const rawEventsForExtrinsic = rawEvents.filter((ev) => ev.phase.isApplyExtrinsic && ev.phase.asApplyExtrinsic.toNumber() === index);
-        return transactionMappers.toPb(index, rawExtrinsic, rawTimestamp, rawEventsForExtrinsic) // todo fees?
+        return transactionMappers.toPb(index, rawExtrinsic, rawTimestamp, rawEventsForExtrinsic, calcFee)
       }),
     }
   };
